# lib/api-zod/src/generated/types/analysisRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * EV Sports Betting Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalysisRequest {
  gameId: string;
  homeTeam: string;
  awayTeam: string;
  sport: string;
  /** @nullable */
  market?: string | null;
  /** @nullable */
  additionalContext?: string | null;
}