# lib/api-zod/src/generated/types/analysisResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * EV Sports Betting Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalysisResult {
  gameId: string;
  analysis: string;
  recommendation: string;
  /** 1-5 star rating */
  confidence: number;
  keyFactors?: string[];
  risks?: string[];
}