# lib/api-zod/src/generated/types/bet.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * EV Sports Betting Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Bet {
  id: number;
  gameId: string;
  homeTeam: string;
  awayTeam: string;
  sport: string;
  market: string;
  selection: string;
  /** @nullable */
  point?: number | null;
  bookmaker: string;
  americanOdds: number;
  evPercent: number;
  units: number;
  /** pending | won | lost | push */
  status: string;
  /** @nullable */
  pnl?: number | null;
  commenceTime: string;
  createdAt: string;
  /** @nullable */
  notes?: string | null;
}