# lib/api-zod/src/generated/types/betStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * EV Sports Betting Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { SportStats } from './sportStats';

export interface BetStats {
  totalBets: number;
  wins: number;
  losses: number;
  pushes: number;
  pending: number;
  roi: number;
  totalUnitsWagered: number;
  totalPnl: number;
  winRate: number;
  bySport?: SportStats[];
}