# lib/api-zod/src/generated/types/betUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * EV Sports Betting Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface BetUpdate {
  /** pending | won | lost | push */
  status?: string;
  /** @nullable */
  pnl?: number | null;
  /** @nullable */
  notes?: string | null;
}